requirements.txt
groq>=0.12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
